refactor: change file name and nomenclature
