--- src/Pages/listOs.tsx
@@ -1,5 +1,0 @@
-const ListOs = () => {
-  return ( <h1> Hello Os</h1> );
-}
- 
-export default ListOs;
